feat(ai-judge): consolidate canonical corpus builder and add temporal trajectories
